+++ backend/src/main/kotlin/io/wury/terra/web/representation/request/GetModsRequest.kt
@@ -1,5 +1,5 @@
 package io.wury.terra.web.representation.request
 
 data class GetModsRequest(
-    val modIds: List<Long>,
+    val modIds: List<Int>,
 )
